requirements.txt
# streamlit
# langchain-openai
# langgraph
# python-dotenv
# vapi_server_sdk
# fastapi
# uvicorn
# vapi_python
# python-multipart


# Frontend dependencies
streamlit
langchain-openai
langgraph
python-dotenv
streamlit-audiorec
requests

# Backend dependencies
fastapi
uvicorn[standard]
python-multipart
pydantic

# Vapi SDK - prova entrambe le versioni
vapi-python>=0.0.1
# oppure se non funziona, usa:
# vapi-server-sdk>=0.1.0